feat: radix switch
