+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@next/font": "^13.0.4",
     "@prisma/client": "^4.5.0",
+    "@radix-ui/react-switch": "^1.0.1",
     "@svgr/webpack": "^6.5.0",
     "@tanstack/react-query": "^4.19.1",
     "@trpc/client": "^10.5.0",
