+++ src/components/global/layout.module.scss
@@ -14,6 +14,7 @@
   width: 100%;
   height: 100vh;
   overflow-y: auto;
+  overflow-x: hidden;
   row-gap: 5px;
 
   @include large {
