+++ src/components/input/switch.tsx
@@ -1,0 +1,47 @@
+import styles from "./switch.module.scss";
+import * as Radix from "@radix-ui/react-switch";
+import React from "react";
+
+const Switch: React.FC<Props> = ({
+  label,
+  name,
+  disabled,
+  value,
+  onChange,
+}: Props) => {
+  const text = label ?? name;
+
+  return (
+    <fieldset
+      className={styles.main}
+      disabled={disabled}
+    >
+      <label className={styles.label} htmlFor={name}>
+        {text}
+      </label>
+
+      <Radix.Root
+        id={name}
+        name={name}
+        checked={value}
+        disabled={disabled}
+        className={styles.root}
+        onCheckedChange={onChange}
+      >
+        <Radix.Thumb className={styles.thumb} />
+      </Radix.Root>
+
+
+    </fieldset>
+  );
+};
+
+export { Switch };
+
+interface Props {
+  name: string;
+  label?: string;
+  disabled?: boolean;
+  value: boolean;
+  onChange: (checked: boolean) => void
+}
